Ignore all the files that aren't necessary for the build.
I've been through all the files I can find trying to work out what's need for the link to succeed.

I don't know that this is a functioning program, it's quite possible that there are some /weak references/ that should be overridden.

This is probably pointless for any real program as the linker shouldn't bring in functions that aren't necessary but I've learnt some stuff while rummaging around.